requirements.txt
streamlit
google-generativeai
requests
pandas
matplotlib